Split Interviews out of the Reviews page into its own pill
/hiring/interviews is the interviewer's home — assigned interviews and the
availability calendar (with Google import), moved verbatim from the reviewer
dashboard's internal ?view= switcher. The Interviews pill gates on a new
isInterviewer role flag (any CycleInterviewer row, mirroring the domain-lead
convention) threaded through every hiring pill call site.

The reviewer page keeps Review + the live Delibs mirror and is now titled
Reviews to match its pill and crumb. The interview detail route moves from
/hiring/interviewer/interview/:id to /hiring/interviews/:id, gains an
applicant-name breadcrumb, and its old prefixes drop out of the breadcrumb
unrouted-segments list. Section panel extracted to a shared component.
