+++ dali-api/app/hiring/routes/domain-lead.tsx
@@ -50,6 +50,7 @@ export async function loader({ request }: Route.LoaderArgs) {
     isCore: roles.isCore,
     isDomainLead: roles.isDomainLead,
     isAdmin: roles.isAdmin,
+    isInterviewer: roles.isInterviewer,
   };
 
   const assignments = await prisma.domainLeadAssignment.findMany({
