+++ dali-api/app/hiring/routes/library.tsx
@@ -57,6 +57,7 @@ export async function loader({ request }: Route.LoaderArgs) {
       isCore: roles.isCore,
       isDomainLead: roles.isDomainLead,
       isAdmin: roles.isAdmin,
+      isInterviewer: roles.isInterviewer,
     },
   };
 }
